# src/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dockore",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/containers": {
      "get": {
        "tags": [
          "containers"
        ],
        "summary": "List Containers",
        "operationId": "list_containers_api_containers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "all",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "All"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContainerItem"
                  },
                  "title": "Response List Containers Api Containers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "containers"
        ],
        "summary": "Create Container",
        "operationId": "create_container_api_containers_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Run"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContainerCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainerItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "containers"
        ],
        "summary": "Delete Containers",
        "operationId": "delete_containers_api_containers_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/containers/{id}": {
      "get": {
        "tags": [
          "containers"
        ],
        "summary": "Get Container",
        "operationId": "get_container_api_containers__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainerItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/containers/{id}/start": {
      "post": {
        "tags": [
          "containers"
        ],
        "summary": "Start Container",
        "operationId": "start_container_api_containers__id__start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/containers/{id}/stop": {
      "post": {
        "tags": [
          "containers"
        ],
        "summary": "Stop Container",
        "operationId": "stop_container_api_containers__id__stop_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "timeout",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Timeout"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/containers/{id}/restart": {
      "post": {
        "tags": [
          "containers"
        ],
        "summary": "Restart Container",
        "operationId": "restart_container_api_containers__id__restart_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "timeout",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Timeout"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/containers/{id}/rename": {
      "post": {
        "tags": [
          "containers"
        ],
        "summary": "Rename Container",
        "operationId": "rename_container_api_containers__id__rename_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/containers/{id}/logs": {
      "get": {
        "tags": [
          "containers"
        ],
        "summary": "Container Logs",
        "operationId": "container_logs_api_containers__id__logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/containers/{id}/diff": {
      "get": {
        "tags": [
          "containers"
        ],
        "summary": "Container Diff",
        "operationId": "container_diff_api_containers__id__diff_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainerDiff"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/containers/{id}/commit": {
      "post": {
        "tags": [
          "containers"
        ],
        "summary": "Commit Container",
        "operationId": "commit_container_api_containers__id__commit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/containers/{id}/exec": {
      "post": {
        "tags": [
          "containers"
        ],
        "summary": "Exec Container",
        "operationId": "exec_container_api_containers__id__exec_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/containers/{id}/terminal": {
      "post": {
        "tags": [
          "containers"
        ],
        "summary": "Create Terminal Ticket Endpoint",
        "operationId": "create_terminal_ticket_endpoint_api_containers__id__terminal_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerminalTicket"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CommitRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CommitRequest"
      },
...
      "ContainerCreate": {
        "properties": {
          "image": {
            "type": "string",
            "title": "Image"
          },
          "command": {
            "type": "string",
            "title": "Command"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "interactive": {
            "type": "boolean",
            "title": "Interactive",
            "default": false
          },
          "tty": {
            "type": "boolean",
            "title": "Tty",
            "default": false
          },
          "ports": {
            "items": {
              "$ref": "#/components/schemas/PortMapping"
            },
            "type": "array",
            "title": "Ports",
            "default": []
          },
          "volumes": {
            "items": {
              "$ref": "#/components/schemas/VolumeMapping"
            },
            "type": "array",
            "title": "Volumes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "image",
          "command"
        ],
        "title": "ContainerCreate"
      },
      "ContainerDiff": {
        "properties": {
          "add": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Add",
            "default": []
          },
          "change": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Change",
            "default": []
          },
          "delete": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Delete",
            "default": []
          },
          "other": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Other",
            "default": []
          }
        },
        "type": "object",
        "title": "ContainerDiff"
      },
      "ContainerItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "image": {
            "$ref": "#/components/schemas/ImageItem"
          },
          "create_time": {
            "type": "string",
            "title": "Create Time"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          },
          "tty": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tty"
          },
          "interactive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interactive"
          },
          "network": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContainerNetwork"
              },
              {
                "type": "null"
              }
            ]
          },
          "mounts": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Mount"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mounts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "image",
          "create_time",
          "status"
        ],
        "title": "ContainerItem"
      },
      "ContainerNetwork": {
        "properties": {
          "ip": {
            "type": "string",
            "title": "Ip",
            "default": ""
          },
          "prefix": {
            "type": "integer",
            "title": "Prefix",
            "default": 0
          },
          "gateway": {
            "type": "string",
            "title": "Gateway",
            "default": ""
          },
          "mac_address": {
            "type": "string",
            "title": "Mac Address",
            "default": ""
          },
          "ports": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PortMapping"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ports"
          }
        },
        "type": "object",
        "title": "ContainerNetwork"
      },
...
      "DeleteResult": {
        "properties": {
          "failed": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Failed",
            "default": {}
          }
        },
        "type": "object",
        "title": "DeleteResult",
        "description": "Failed ids mapped to their error message; empty means fully succeeded."
      },
...
      "ExecRequest": {
        "properties": {
          "command": {
            "type": "string",
            "title": "Command"
          },
          "interactive": {
            "type": "boolean",
            "title": "Interactive",
            "default": true
          },
          "tty": {
            "type": "boolean",
            "title": "Tty",
            "default": true
          },
          "privileged": {
            "type": "boolean",
            "title": "Privileged",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "command"
        ],
        "title": "ExecRequest"
      },
      "ExecResult": {
        "properties": {
          "exit_code": {
            "type": "integer",
            "title": "Exit Code"
          },
          "output": {
            "type": "string",
            "title": "Output"
          }
        },
        "type": "object",
        "required": [
          "exit_code",
          "output"
        ],
        "title": "ExecResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IdsRequest": {
        "properties": {
          "ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Ids"
          }
        },
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "IdsRequest"
      },
      "ImageItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "author": {
            "type": "string",
            "title": "Author",
            "default": ""
          },
          "create_time": {
            "type": "string",
            "title": "Create Time"
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "default": 0
          },
          "command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          },
          "tty": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tty"
          },
          "interactive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interactive"
          },
          "architecture": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Architecture"
          },
          "os": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Os"
          },
          "ports": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ExposedPort"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ports"
          }
        },
        "type": "object",
        "required": [
          "id",
          "create_time"
        ],
        "title": "ImageItem"
      },
...
      "Mount": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "driver": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": ""
          },
          "src": {
            "type": "string",
            "title": "Src",
            "default": ""
          },
          "dest": {
            "type": "string",
            "title": "Dest",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "Mount"
      },
...
      "RenameRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RenameRequest"
      },
...
      "StatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "StatusResponse"
      },
...
      "TerminalRequest": {
        "properties": {
          "command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          }
        },
        "type": "object",
        "title": "TerminalRequest"
      },
      "TerminalTicket": {
        "properties": {
          "ticket": {
            "type": "string",
            "title": "Ticket"
          },
          "expires": {
            "type": "integer",
            "title": "Expires"
          }
        },
        "type": "object",
        "required": [
          "ticket",
          "expires"
        ],
        "title": "TerminalTicket"
      },
...
      "VolumeMapping": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "bind": {
            "type": "string",
            "title": "Bind"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "rw"
          }
        },
        "type": "object",
        "required": [
          "path",
          "bind"
        ],
        "title": "VolumeMapping"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}